+++ changelog.md
@@ -6,5 +6,7 @@
 - Added a rockspec and documented luarocks dependencies
 
 ## [0.1.1] - 2026-01-30
-### Docs
+### Features
+- Added persistent sessions (on by default, can be toggled using `:TungstenTogglePersistence`)
+### documentation
 - Updated documentation index to include all the relevant information from the README ontop of the previous content